+++ bareops-error/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "bareops-error"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+miette = "7.5.0"
+thiserror = "2.0.12"
