+++ bareops-inventory/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "bareops-inventory"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
