+++ bareops-inventory/src/lib.rs
@@ -1,0 +1,1 @@
+pub struct Target {}
